Add files via upload
This script performs PCA, UMAP, and t-SNE on gene expression data, analyzing samples individually and combined. To use, place the CSV file in the working directory, adjust file paths as needed, and ensure tidyverse, umap, and Rtsne packages are installed. Run the script in an R environment for PCA and dimensionality reduction visualizations. NEED TO CHECK THIS, as this version (for a specific project) does not match my loop version (also uploaded as "w group loops") and I'm not completely sure why yet.

Make sure to pay attention to inputs like perplexity and number of neighbors, as those may depend on your sample size