Add files via upload
This script performs PCA, UMAP, and t-SNE on gene expression data, analyzing samples individually and combined. To use, place the CSV file in the working directory, adjust file paths as needed, and ensure tidyverse, umap, and Rtsne packages are installed. 

This script is based off of the previous R PCA UMAP tSNE (also uploaded to repository) but I tried to implement loops so that the user could edit group # and names. This updated script tried to adjust n_neighbors for UMAP and perplexity for t-SNE based on sample size, but for now some of those are auto placed so definitely edit as needed.